\name{analysis}
\alias{analysis}
\alias{loglogAnalysis}
\title{Analysis functions for GPU/CPU Benchmarking}
\description{
  Analysis functions for GPU/CPU Benchmarking
}
\usage{
  loglogAnalysis()
}
\details{
  \code{loglogAnalysis} retrieves past benchmark results from the
  database contained in the package and returns intercepts and slops of
  regressions of elapsed times on matrix dimensions (where both inputs
  are in logarithms).
}  
\keyword{misc}
